Use relation stats extended instead of column stats extended
