+++ input/postgres/schema.go
@@ -95,10 +95,10 @@ func collectOneSchema(ctx context.Context, server *state.Server, collectionOpts 
 	}
 
 	ps.SchemaStats[databaseOid] = &state.SchemaStats{
-		RelationStats:       make(state.PostgresRelationStatsMap),
-		IndexStats:          make(state.PostgresIndexStatsMap),
-		ColumnStats:         make(state.PostgresColumnStatsMap),
-		ColumnStatsExtended: make(state.PostgresColumnStatsExtendedMap),
+		RelationStats:         make(state.PostgresRelationStatsMap),
+		IndexStats:            make(state.PostgresIndexStatsMap),
+		ColumnStats:           make(state.PostgresColumnStatsMap),
+		RelationStatsExtended: make(state.PostgresRelationStatsExtendedMap),
 	}
 
 	psOut, tsOut, err = collectSchemaData(ctx, collectionOpts, logger, schemaConnection, ps, ts, databaseOid, postgresVersion, server, systemType, dbName)
@@ -141,12 +141,12 @@ func collectSchemaData(ctx context.Context, collectionOpts state.CollectionOpts,
 			ps.SchemaStats[databaseOid].ColumnStats[k] = v
 		}
 
-		newColumnStatsExtended, err := GetColumnStatsExtended(ctx, logger, db, postgresVersion, server, collectionOpts, systemType, dbName)
+		newRelationStatsExtended, err := GetRelationStatsExtended(ctx, logger, db, postgresVersion, server, collectionOpts, systemType, dbName)
 		if err != nil {
 			return ps, ts, fmt.Errorf("error collecting extended column statistics: %s", err)
 		}
-		for k, v := range newColumnStatsExtended {
-			ps.SchemaStats[databaseOid].ColumnStatsExtended[k] = v
+		for k, v := range newRelationStatsExtended {
+			ps.SchemaStats[databaseOid].RelationStatsExtended[k] = v
 		}
 	}
